+++ lib/json_statham/result.rb
@@ -1,8 +1,11 @@
 # frozen_string_literal: true
 
 module JsonStatham
+  class InvalidRatioError < StandardError
+  end
+
   class Result
-    attr_reader :parser
+    attr_reader :parser, :config
 
     def self.call(parser)
       new(parser).call
@@ -12,15 +15,11 @@ module JsonStatham
       Validation.check_object_class(parser, [JsonStatham::Parser])
 
       @parser = parser
+      @config = JsonStatham.config
     end
 
     def call
-      if JsonStatham.config.logger?
-        puts <<-RESULT_MSG
-          Previous duration: #{previous_duration}
-          Current duration: #{current_duration}
-        RESULT_MSG
-      end
+      ensure_valid_ratio
 
       self
     end
@@ -39,12 +38,34 @@ module JsonStatham
       parser.previous_duration
     end
 
+    def previous_duration?
+      !!previous_duration
+    end
+
     def success?
       parser.valid?
     end
 
     def failure?
       !success?
+    end
+
+    def ratio
+      return 0 unless observed?
+
+      current_duration.fdiv(previous_duration)
+    end
+
+    def raise_ratio
+      config.raise_ratio
+    end
+
+    private
+
+    def ensure_valid_ratio
+      return unless config.raise_on_failure? && previous_duration? && ratio > raise_ratio
+
+      raise JsonStatham::InvalidRatioError, "raise_ratio: #{raise_ratio}. current_ratio: #{ratio}"
     end
   end
 end