+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### [0.1.1] - 2023-02-07
+
+  * Features:
+    * Add `raise_ratio` to config.
+
+  * Deprecations:
+    * Remove `logger` from config.
+
 ### [0.1.0] - 2023-02-06
 
 - Initial release